add frame for accounts, orders, authentications, products

+++ src/main/java/com/graduate/hou/dto/DailyRevenueDTO.java
@@ -22,5 +22,5 @@ public class DailyRevenueDTO {
 
     private BigDecimal totalRevenue;
 
-    private Integer totalOrders;
+    private Long totalOrders;
 }
